refs #32860 clean up muon GUI grouping code
