+++ qt/python/mantidqtinterfaces/mantidqtinterfaces/Muon/GUI/Common/ADSHandler/workspace_naming.py
@@ -10,8 +10,8 @@ GROUP_STR = "; Group; "
 DIFF_STR = "; Diff; "
 PAIR_STR = "; Pair Asym; "
 PHASEQUAD_STR = '; PhaseQuad; '
-PHASEQUAD_RE = ' Re ' # use _ on both sides to prevent accidental ID (e.g. Red)
-PHASEQUAD_IM = " Im "
+PHASEQUAD_RE = '_Re_' # use _ on both sides to prevent accidental ID (e.g. Red)
+PHASEQUAD_IM = "_Im_"
 TF_ASYMMETRY_PREFIX = "TFAsymmetry"
 REBIN_STR = 'Rebin'
 FFT_STR = 'FFT'
