+++ qt/python/mantidqtinterfaces/mantidqtinterfaces/Muon/GUI/Common/utilities/algorithm_utils.py
@@ -15,15 +15,6 @@ from mantid.kernel import PhysicalConstants as const
 muon_logger = Logger('Muon-Algs')
 
 
-def make_group(ws_list, group_name):
-    alg = mantid.AlgorithmManager.create("GroupWorkspaces")
-    alg.initialize()
-    alg.setAlwaysStoreInADS(True)
-    alg.setProperty("InputWorkspaces", ws_list)
-    alg.setProperty("OutputWorkspace", group_name)
-    alg.execute()
-
-
 def create_empty_table(name):
     alg = mantid.AlgorithmManager.create("CreateEmptyTableWorkspace")
     alg.initialize()
